+++ src/test/kotlin/org/jesperancinha/plugins/omni/reporter/utils/Utils.kt
@@ -2,6 +2,8 @@ package org.jesperancinha.plugins.omni.reporter.utils
 
 import java.io.File
 import java.io.FileNotFoundException
+import java.net.URL
+import kotlin.io.path.toPath
 
 class Utils {
     companion object {
@@ -10,3 +12,5 @@ class Utils {
         )
     }
 }
+
+fun URL.toFile(): File = this.toURI().toPath().toFile()