+++ ReleaseNotes.md
@@ -3,8 +3,14 @@
 
 #### Release 0.1.5 - Upcoming
 
+###### Features
+
 1. Support for Group Coverage with Branch Coverage (mostly Coveralls)
 2. Source encoding gets automatically chosen unless we configure flag `failOnNoEncoding` to `true`
+
+###### Bugs
+
+1. Jacoco XML generation sometimes results in empty string.
 
 #### Release 0.1.4 - 2022/02/1